+++ db/migrate/20170113163002_create_pedals.rb
@@ -2,8 +2,8 @@ class CreatePedals < ActiveRecord::Migration
   def change
     create_table :pedals do |t|
       t.string :name, null: false, index: { unique: true }
-      t.decimal :length
-      t.decimal :width
+      t.float :length
+      t.float :width
 
       t.timestamps null: false
     end
